JMFCameraIOS v0.38 Changed Face Detection View Controller

+++ JMFCameraIOS_AlbumViewController.m
@@ -31,7 +31,7 @@
 /***************************************************************************/
 @interface JMFCameraIOS_AlbumViewController ()
 {
-    int     albumCount;
+    long     albumCount;
 }
 
 @end
